Move HomeHero out of shared components dir into Home dir
