+++ apps/web/src/pages/Home/HomePage/components/Hero/Hero.tsx
@@ -5,7 +5,6 @@ import MagnifyingGlassIcon from "@heroicons/react/24/outline/MagnifyingGlassIcon
 import { Heading, Link } from "@gc-digital-talent/ui";
 import { commonMessages, navigationMessages } from "@gc-digital-talent/i18n";
 
-import HomeHero from "~/components/HeroDeprecated/HomeHero";
 import useRoutes from "~/hooks/useRoutes";
 import hero1Landscape from "~/assets/img/hero-1-landscape.webp";
 import hero1Portrait from "~/assets/img/hero-1-portrait.webp";
@@ -19,6 +18,8 @@ import hero3Tablet from "~/assets/img/hero-3-tablet-portrait.webp";
 import hero4Landscape from "~/assets/img/hero-4-landscape.webp";
 import hero4Portrait from "~/assets/img/hero-4-portrait.webp";
 import hero4Tablet from "~/assets/img/hero-4-tablet-portrait.webp";
+
+import HomeHero from "../../../components/HomeHero";
 
 const landscapeRandomize = (index?: number) => {
   const imageSets = {
