+++ apps/web/src/pages/Home/components/HomeHero.tsx
